# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/problems": {
      "post": {
        "description": "Cria um novo problema de infraestrutura",
        "operationId": "CreateProblemController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProblemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Problema criado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Token JWT inválido ou expirado"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Criar problema",
        "tags": ["Problems"]
      },
      "get": {
        "description": "Retorna uma lista paginada de problemas reportados no sistema com informações de localização",
        "operationId": "FetchProblemsController_handle",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página (começa em 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Quantidade de itens por página",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Incluir problemas deletados (na lixeira)",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "statuses",
            "required": false,
            "in": "query",
            "description": "Lista de status separada por vírgula para filtrar problemas",
            "schema": {
              "example": "TO_ANALYSIS,IN_PROGRESS",
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Termo de busca para filtrar problemas por título ou descrição",
            "schema": {
              "example": "ar condicionado",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de problemas retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "problems": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProblemWithDetailsResponse"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 42
                    },
                    "page": {
                      "type": "number",
                      "example": 1
                    },
                    "pageSize": {
                      "type": "number",
                      "example": 20
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros inválidos"
          }
        },
        "summary": "Buscar problemas",
        "tags": ["Problems"]
      }
    },
...
    "/categories": {
      "get": {
        "description": "Retorna uma lista paginada de categorias de problemas disponíveis no sistema",
        "operationId": "FetchCategoriesController_handle",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página (começa em 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Quantidade de itens por página",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Incluir categorias deletadas (na lixeira)",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrar por status ativo/inativo",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Busca por nome ou descrição",
            "schema": {
              "example": "Climatização",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorias retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "categories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CategoryResponse"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 42
                    },
                    "page": {
                      "type": "number",
                      "example": 1
                    },
                    "pageSize": {
                      "type": "number",
                      "example": 20
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros inválidos"
          },
          "403": {
            "description": "Usuário não tem permissão para listar itens deletados"
          }
        },
        "summary": "Buscar categorias",
        "tags": ["Categories"]
      },
      "post": {
        "description": "Cria uma nova categoria de problema (requer role MANAGER+)",
        "operationId": "CreateCategoryController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoria criada com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Token JWT inválido ou expirado"
          },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Criar categoria",
        "tags": ["Categories"]
      }
    },
...
    "/locations": {
      "post": {
        "description": "Cria uma nova localização no sistema (requer role MANAGER+)",
        "operationId": "CreateLocationController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Localização criada com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Token JWT inválido ou expirado"
          },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Criar localização",
        "tags": ["Locations"]
      },
      "get": {
        "description": "Retorna uma lista paginada de localizações disponíveis no sistema",
        "operationId": "FetchLocationsController_handle",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página (começa em 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Quantidade de itens por página",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Incluir localizações deletadas (na lixeira)",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrar por status ativo/inativo",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Termo de busca para filtrar localizações por nome, código ou descrição",
            "schema": {
              "example": "Bloco A",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de localizações retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "locations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LocationResponse"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 42
                    },
                    "page": {
                      "type": "number",
                      "example": 1
                    },
                    "pageSize": {
                      "type": "number",
                      "example": 20
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros inválidos"
          },
          "403": {
            "description": "Usuário não tem permissão para listar itens deletados"
          }
        },
        "summary": "Buscar localizações",
        "tags": ["Locations"]
      }
    },
...
  },
  "info": {
    "title": "IFAL Arapiraca Backend API",
    "description": "API para gerenciamento de problemas de infraestrutura do IFAL Arapiraca",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "CreateProblemRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Título do problema",
            "example": "Ar condicionado não está funcionando"
          },
          "description": {
            "type": "string",
            "description": "Descrição detalhada do problema",
            "example": "O ar condicionado da sala 201 não está ligando há 3 dias"
          },
          "categoryId": {
            "type": "string",
            "description": "ID da categoria do problema",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "locationId": {
            "type": "string",
            "description": "ID da localização onde ocorreu o problema",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "attachmentIds": {
            "description": "IDs dos anexos (retornados pelo upload)",
            "example": ["123e4567-e89b-12d3-a456-426614174002"],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["title", "description", "categoryId", "locationId"]
      },
      "ProblemReporterResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do autor do problema",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "email": {
            "type": "string",
            "description": "Email do autor do problema",
            "example": "joao.silva@ifal.edu.br"
          }
        },
        "required": ["id", "email"]
      },
      "AttachmentInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do anexo",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "title": {
            "type": "string",
            "description": "Título do anexo",
            "example": "foto-ar-condicionado.jpg"
          },
          "url": {
            "type": "string",
            "description": "URL do anexo",
            "example": "https://i.ibb.co/xxxxx/foto-ar-condicionado.jpg"
          }
        },
        "required": ["id", "title", "url"]
      },
      "ProblemWithDetailsResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do problema",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "reporter": {
            "description": "Dados do autor do problema",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProblemReporterResponse"
              }
            ]
          },
          "title": {
            "type": "string",
            "description": "Título do problema",
            "example": "Ar condicionado não funcionando"
          },
          "slug": {
            "type": "string",
            "description": "Slug do problema",
            "example": "ar-condicionado-nao-funcionando"
          },
          "excerpt": {
            "type": "string",
            "description": "Trecho da descrição do problema (primeiros 120 caracteres)",
            "example": "O ar condicionado da sala 201 não está ligando há 3 dias..."
          },
          "locationName": {
            "type": "string",
            "description": "Nome da localização do problema",
            "example": "Bloco A - Sala 201"
          },
          "status": {
            "type": "string",
            "description": "Status do problema",
            "enum": [
              "TO_ANALYSIS",
              "IN_ANALYSIS",
              "ACCEPTED",
              "REJECTED",
              "IN_PROGRESS",
              "FINISHED"
            ],
            "example": "TO_ANALYSIS"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de criação",
            "example": "2025-01-15T10:30:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de atualização",
            "example": "2025-01-15T10:30:00Z"
          },
          "deletedAt": {
            "type": "object",
            "description": "Data de exclusão (null se não deletado)",
            "example": "2025-01-15T10:30:00Z",
            "nullable": true
          },
          "attachments": {
            "description": "Anexos do problema",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentInfo"
            }
          }
        },
        "required": [
          "id",
          "reporter",
          "title",
          "slug",
          "excerpt",
          "locationName",
          "status",
          "createdAt"
        ]
      },
...
      "CreateCategoryRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da categoria",
            "example": "Climatização"
          },
          "description": {
            "type": "string",
            "description": "Descrição da categoria",
            "example": "Problemas relacionados a ar condicionado, ventilação e climatização"
          }
        },
        "required": ["name", "description"]
      },
...
      "CreateLocationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da localização",
            "example": "Bloco A - Sala 201"
          },
          "description": {
            "type": "string",
            "description": "Descrição da localização",
            "example": "Sala de aula localizada no primeiro andar do Bloco A"
          },
          "code": {
            "type": "string",
            "description": "Código identificador da localização",
            "example": "BLA-201"
          }
        },
        "required": ["name", "description", "code"]
      },
...
    }
  }
}